gh-93649: Split exception tests from _testcapimodule.c
